# ENH: improve performance of lexsort for integers.