This PR implements an alternative and more performant code path for ``lexsort`` when passed (appropriate - see below) integer ``keys``.
It does so by compressing the levels into a single array using bit shifts, and then sorting the resulting array.

The result of this is, at a glance (details below):
- ~1.75x speedup for large arrays (N>2000)
- ... but a much large speedup (e.g. 8x in my application) can be obtained thanks to the possibility of selecting the desired sorting algorithm
- ~0.75x slowdown for small arrays (N<1000)
- very small slowdown for the (rare, I think) case where ``keys`` are integers but they are too large for the new code path (as reported in the docs, this is guaranteed not to happen for instance if all values are below MAX_INT / 2L, L being the number of levels)

The application I'm referring to is [the use of ``lexsort`` inside COO sparse matrices in scipy](https://github.com/scipy/scipy/issues/20151).

The current PR was influenced by some feedback received by @ngoldbaum [in the mailing list](https://mail.python.org/archives/list/numpy-discussion@python.org/thread/Q5EH4NJU3HEHWL6YB46KRZASRXYII2HM/). However, I did not use C, because the Python code I wrote is vectorized anyway, and writing the same code in C would take me too much time (my C skills being full of rust) resulting in less readable code.

Assuming that, despite this, the approach is of interest for numpy (if this PR is not, I guess I will make an attempt directly in scipy?),

- [ ]  I must obviously solve the problem of where to position the Python code (I don't think the current file is right, since it contains lower level code and cleanly introducing the imports I need would result in a circular import - this is why there currently is a horrible import inside the function): guidance on this will be highly appreciated
- [ ]  I must also understand why at the end of the function I need to make ``axis`` explicit, but this is related to the previous point I think
- [ ] I must ~write tests~ integrate the tests I wrote into the codebase

Few other things to note:
- the approach is similar to [the one I took](https://github.com/pandas-dev/pandas/pull/19074) in the ``MultiIndex`` engine for ``pandas``
- a previous version of this PR (before feedback from @ngoldbaum ) would allow the user to switch on or off the new code path, but indeed the checks required to establish if it is feasible add a negligible overhead
- I *think* that it would be easy to patch [the C code](https://github.com/numpy/numpy/blob/main/numpy/_core/src/multiarray/item_selection.c#L1796) as to support different sorting algorithms... at least easy for someone a bit more fluent in C, definitely not for me. If such a patch is introduced, this PR will not guarantee the 8x speedup on nearly-sorted arrays, just the ~1.75x speedup.

Below are the result of the performance tests, where
- "PR" is this branch, "base" is main (previous commit)
- "mix" refers to random arrays, "sep" to arrays where two sorted (2-level) arrays are concatenated (so on which ``stable`` sort shines, because it has linear cost)
- "sta" refers to tests with ``kind=sep`` (in the PR)
- "big" refers to arrays where values are so large that the new code path cannot be taken (and a minimum of overhead is wasted in order to determine this is the case)


Absolute scale:
![abs](https://github.com/numpy/numpy/assets/1224492/3b9a159b-bd64-49f5-bc6b-2e25b602a755)

Log scale:
![log](https://github.com/numpy/numpy/assets/1224492/97f8530e-a4bc-4a59-91c1-95f79062d67b)

Estimation of speedup ratio when the new code path is feasible:

![speedup](https://github.com/numpy/numpy/assets/1224492/c18595e8-ea90-4833-b8a2-2327a3fba4cf)

... and when it is not (note that it is never taken for N < 1000, so the loss in efficiency is overhead):

![speedup_big](https://github.com/numpy/numpy/assets/1224492/51ea3ec3-5839-4621-acbd-deb9eabd6dbe)

The code used for these tests is available [here](https://github.com/toobaz/int_lexsort/blob/main/Show%20benchmarks%20-%20PR.ipynb).