@toobaz 2024-05-14
I'm confused... the failing test is
```
         v = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10])
         assert_equal(np.lexsort(v), 0)
```
... but the docs state ``Each element of `keys` along the zeroth axis must be an array-like object of the same shape.`` So why are we checking the result - rather than expecting an error - in a case where ``keys`` is 1D (that is, elements of ``keys`` are scalar)?!