@ngoldbaum 2024-05-14
Scalars are a kind of array-like. My reading of the docs is this case is allowed, with `N=0`. The test being triggered is [18 years old](https://github.com/numpy/numpy/commit/9a0c3ebd59782e352b3fc6357577751065ff691f) so this is behavior that's been stable for the entire life of the library it looks like.