@ngoldbaum 2024-05-15
We looked at this at the triage meeting. Overall we think this is a neat idea but there are some issues.

First, can you add a one or two line comment above the first check that dispatches to the fast integer path explaining exactly what the fast path is doing. It would also be nice if you could link to an explanation of how the compression works, particularly if you can find any discussions about doing that for a sorting algorithm.

Do we really need the new `kind` argument? It's not great IMO to add a keyword argument which only applies based on particular properties of the input data.

We also found the coding style to be a little bit opaque. Can you try to break up some of the longer lines that combine many operations to make the algorithm a little clearer?

This is sort of on the bubble for the sort of contribution we'd take that improves performance. We have to balance maintainability too. Improvements that make this code easier to understand would tilt the balance towards merging this.