@ngoldbaum 2024-05-15
Oh also it would be nice if you could add a new benchmark for `lexsort` to the existing benchmark suite and show that it improves performance for the case you're looking for and also that the performance hit is negligible for the un-optimized case. Ping here if you have trouble with the benchmark suite, happy to answer questions.