@toobaz 2024-05-21
Again thanks for your great feedback @ngoldbaum 

I will soon edit my PR accordingly, but first let me take a step back:

> Do we really need the new `kind` argument? It's not great IMO to add a keyword argument which only applies based on particular properties of the input data.

Indeed, we do not really need it in this PR (at least for my purposes, since the default "stable" is precisely what I need to get the huge efficiency gain!).

This said, I decided to take a look at whether in fact supporting ``kind`` in the current code path was maybe a low hanging fruit.

Turned out it isn't (at least not low enough for my awful C skills). My understanding is that ``lexsort`` basically calls a normal ``mergesort`` on a rearranged version of the input keys. So I tried to just replace mergesort with the other ``kind``s there... and I failed.

The attempt is here: https://github.com/numpy/numpy/compare/main...toobaz:numpy:lexsort_kind

... but it fails brutally, already with
``np.lexsort([[2, 7], [0, 0]], kind='h')`` which returns ``[1, 0]``
(while ``np.lexsort([[2, 7], [0, 0]])``, ``np.argsort([2, 7], kind='h')`` and ``np.lexsort([[2, 7], [0, 0]], kind='s')`` all correctly return ``[0, 1]``).

If there's any chance someone good at C can maybe identify some stupid mistake I made, so that we get ``kind`` supported in general, great. Otherwise, if either the problem is deeper, or anyway nobody has time to check my (other) code, I will just remove support for ``kind`` from this PR.