@toobaz 2024-06-03
OK, never mind (at least in this PR) the ``kind`` argument.

This new version drops it, makes code way more readable (I hope), and adds tests and benchmarks, here's the result:

```
| Change   | Before [76807f0e] <main>   | After [94dea95c] <int_lexsort_kind>   |   Ratio | Benchmark (Parameter)                                                      |
|----------|----------------------------|---------------------------------------|---------|----------------------------------------------------------------------------|
| +        | 1.70±0.02μs                | 2.07±0.02μs                           |    1.22 | bench_function_base.LexSort.time_lexsort('int64', 2, 100, 'mixed')         |
| +        | 1.81±0.03μs                | 2.20±0.02μs                           |    1.22 | bench_function_base.LexSort.time_lexsort('int64', 2, 100, 'separated')     |
| +        | 1.74±0.02μs                | 2.10±0.02μs                           |    1.21 | bench_function_base.LexSort.time_lexsort('int64', 2, 100, 'big')           |
| +        | 3.06±0.01μs                | 3.51±0.02μs                           |    1.15 | bench_function_base.LexSort.time_lexsort('int64', 4, 100, 'separated')     |
| +        | 2.38±0.08μs                | 2.71±0.01μs                           |    1.14 | bench_function_base.LexSort.time_lexsort('float64', 2, 100, 'mixed')       |
| +        | 2.50±0.06μs                | 2.85±0.01μs                           |    1.14 | bench_function_base.LexSort.time_lexsort('float64', 2, 100, 'separated')   |
| +        | 2.93±0.03μs                | 3.34±0.02μs                           |    1.14 | bench_function_base.LexSort.time_lexsort('int64', 4, 100, 'mixed')         |
| +        | 2.46±0.08μs                | 2.79±0.01μs                           |    1.13 | bench_function_base.LexSort.time_lexsort('float64', 2, 100, 'big')         |
| +        | 2.99±0.03μs                | 3.38±0.02μs                           |    1.13 | bench_function_base.LexSort.time_lexsort('int64', 4, 100, 'big')           |
| +        | 4.11±0.03μs                | 4.61±0.04μs                           |    1.12 | bench_function_base.LexSort.time_lexsort('float64', 4, 100, 'mixed')       |
| +        | 4.20±0.1μs                 | 4.67±0.03μs                           |    1.11 | bench_function_base.LexSort.time_lexsort('float64', 4, 100, 'big')         |
| -        | 12.6±0.03ms                | 7.12±0.06ms                           |    0.56 | bench_function_base.LexSort.time_lexsort('int64', 2, 100000, 'mixed')      |
| -        | 905±2μs                    | 490±5μs                               |    0.54 | bench_function_base.LexSort.time_lexsort('int64', 2, 10000, 'mixed')       |
| -        | 173±1ms                    | 92.8±0.8ms                            |    0.54 | bench_function_base.LexSort.time_lexsort('int64', 2, 1000000, 'mixed')     |
| -        | 171±2ms                    | 18.0±0.2ms                            |    0.11 | bench_function_base.LexSort.time_lexsort('int64', 2, 1000000, 'separated') |
| -        | 907±3μs                    | 85.7±1μs                              |    0.09 | bench_function_base.LexSort.time_lexsort('int64', 2, 10000, 'separated')   |
| -        | 12.7±0.1ms                 | 962±30μs                              |    0.08 | bench_function_base.LexSort.time_lexsort('int64', 2, 100000, 'separated')  |
```

As expected, small overhead costs on small arrays, large gains on large arrays.

@ngoldbaum I'm pretty sure that the code is still in the wrong file (forcing me to ``import numpy as np`` within the function) and that (probably relatedly) I'm misusing the "dispatcher"; but I don't understand your suggestion to rename ``_general_lexsort`` to ``lexsort`` - the latter is the name of the API method, which is now a Python method without corresponding C method. Should I rename the C method instead?

This said, there are two types of errors in the tests.
- (some?) 32 bit platforms give wrong results, and I will investigate this as soon as I can find a test machine (I guess a Raspberry Pi will do)
- the ``warnings.simplefilter("ignore", RuntimeWarning)`` at line 610 breaks some assertion... and on this I need help, because I really don't understand the error. It seems to refer to a warning being raised, but in my case a warning is being ignored!